Generate thumbnails lazily in My Files grid + list views
